Refactor some common code out to new file
